requirements.txt
mysql-connector-python
PyQt6
pyqt6-tools
PyYAML
types-PyYAML
